+++ CHANGELOG.md
@@ -4,18 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 8193.32-HEAD
-https://github.com/nwnxee/unified/compare/build8193.32...HEAD
+## 8193.33-HEAD
+https://github.com/nwnxee/unified/compare/build8193.33...HEAD
 
 ### Added
-- Events: added event `NWNX_ON_CLIENT_SET_DEVICE_PROPERTY_BEFORE__*` which fires when a player changes a device property (window size/gui scale).
-- Tweaks: added `NWNX_TWEAKS_SUMMON_ASSOCIATE_USE_COMPUTE_SAFE_LOCATION` which makes CNWSCreature::SummonAssociate() will use ComputeSafeLocation() instead of ComputeNonVisibleLocation(), possibly fixing some Familiar/AnimalCompanion summon issues.
+- N/A
 
 ##### New Plugins
 - N/A
 
 ##### New NWScript Functions
-- Util: GetModuleMtime()
+- N/A
 
 ### Changed
 - N/A
@@ -25,6 +24,19 @@ https://github.com/nwnxee/unified/compare/build8193.32...HEAD
 
 ### Removed
 - N/A
+
+### Fixed
+- N/A
+
+## 8193.33
+https://github.com/nwnxee/unified/compare/build8193.32...build8193.33
+
+### Added
+- Events: added event `NWNX_ON_CLIENT_SET_DEVICE_PROPERTY_BEFORE__*` which fires when a player changes a device property (window size/gui scale).
+- Tweaks: added `NWNX_TWEAKS_SUMMON_ASSOCIATE_USE_COMPUTE_SAFE_LOCATION` which makes CNWSCreature::SummonAssociate() will use ComputeSafeLocation() instead of ComputeNonVisibleLocation(), possibly fixing some Familiar/AnimalCompanion summon issues.
+
+##### New NWScript Functions
+- Util: GetModuleMtime()
 
 ### Fixed
 - Chat: When using custom hearing distances, broadcast the actual message instead of the tellname.